# .harness/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 백주대간 AI 매칭·탐색 서비스 API
  version: 0.1.0-contract
  description: >
    상태: CONFIRMED (CTR-002). 이 문서는 다음을 근거로 작성됐다:
    docs/redesign-v2/web/W-8-api-and-db-changes.md §5, docs/redesign-v2/common/
    C-5-common-api.md §1, 기존 backend/app/api/v1/endpoints/{ontology,system}.py,
    backend/app/api/health.py(BAC-001 동시 구현 중 - 이 문서는 그 계약만 문서화하고
    backend/app/api/**는 직접 수정하지 않는다), WAVE-1 §7.4 엔드포인트 목록.
    CR-001(2026-08-02) 이후 채널 전략은 WEB_ONLY로 변경 적용되어 전용 KIOSK
    채널은 GUEST_WEB 모바일 웹 세션으로 대체한다. 기존 kiosk 구현/마이그레이션은
    즉시 삭제하지 않고 deprecated 후 cleanup migration으로 처리한다.
    스키마 이름은 .harness/contracts/domain-model.md §13/§15 크로스워크를 따라 실제
    존재하는 스키마.테이블명을 반영한다(프롬프트의 평평한 테이블명이 아니라).
    오류 응답은 .harness/contracts/error-codes.yaml, 온톨로지 코드는
    .harness/contracts/ontology.yaml을 따른다.
  contact:
    name: CONTRACTS track
    url: "https://github.com/example/backju/.harness/contracts"
servers:
  - url: /api/v1
    description: 버전 프리픽스 (backend/app/core/config.py API_V1_PREFIX)
...
security:
  - bearerAuth: []
paths:
...
  /exhibitors/{exhibitor_id}:
    get:
      operationId: getExhibitor
      summary: 업체 상세 조회 (승인된 업체만 - EXHIBITOR_NOT_PUBLISHED)
      tags: [exhibitors]
      security: []
      parameters:
        - name: exhibitor_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 업체 상세
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExhibitorDetailResponse" }
        "404":
          description: 존재하지 않거나 미승인(master_approval_status != APPROVED, AGENTS.md §8)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /booths/{booth_id}:
    get:
      operationId: getBooth
      summary: 부스 상세 조회 (운영상태·혼잡도 포함)
      tags: [exhibitors]
      security: []
      parameters:
        - name: booth_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 부스 상세
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoothDetailResponse" }
        "404":
          $ref: "#/components/responses/NotFound"

  # ==========================================================================
  # SEARCH (W-8 §5 POST /search + C-5 §1 POST /common/search, matching.search_*)
  # CR-001 이후 REGISTERED_WEB/GUEST_WEB/BUYER_WEB이 같은 검색 파사드를 공유한다.
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        profile.user_account 인증 세션 토큰. GUEST_WEB 익명 경로는 security: []로
        명시된 엔드포인트에서만 guest_session_id(바디/경로)와 단기 토큰으로 대체 인가한다.
...
  responses:
    NotFound:
      description: RESOURCE_NOT_FOUND
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: ".harness/contracts/error-codes.yaml http_error_codes[].code 중 하나"
              example: VALIDATION_ERROR
            message: { type: string }
            request_id: { type: string, format: uuid }
            details:
              type: object
              nullable: true
              additionalProperties: true
...
    ExhibitorDetailResponse:
      type: object
      properties:
        exhibitor_id: { type: string, format: uuid }
        exhibitor_name: { type: string }
        business_types:
          type: array
          items: { type: string, description: "온톨로지 코드" }
        booths:
          type: array
          items: { type: string, format: uuid }
    BoothDetailResponse:
      type: object
      properties:
        booth_id: { type: string, format: uuid }
        exhibitor_id: { type: string, format: uuid }
        operating_status: { type: string, enum: [OPEN, PAUSED, CLOSED] }
        congestion_level: { type: string, enum: [LOW, MEDIUM, HIGH, UNKNOWN] }
        row_version: { type: integer }

    # ------------------------------------------------------------------------
    # SEARCH (matching.search_session/query/result - CTR-008)
    # ------------------------------------------------------------------------
...